Add files via upload
A program to find whether a given string is a comment or not. If its a comment, print out whether its a single line comment or multiple line comment. Take multiple input from user.

Sample input:
//All is well.
/*My name is John./
/*I love EDU.*/
/#Long live Bangladesh.
Sample output:
Single line comment.
Not a comment.
Multiple line comment.
Not a comment.